Create Package DTO

+++ BeanStyle-Hub/src/main/java/com/beanjoseph/BeanStyle_Hub/model/Material.java
@@ -12,23 +12,23 @@ public class Material {
 
     }
 
-    public Material(Long id, String name) {
+    public Material(Integer id, String name) {
         this.id = id;
         this.name = name;
     }
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     @Column(name = "name")
     private String name;
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
